feat(financial): harden numeric representation
Carry money, price and volume as exact Decimal in every provider/domain
contract and in portfolio aggregation, converting MT5 floats at the provider
boundary with Decimal(str(...)) so no binary-float artifact enters the
domain. Ratios and tick counts stay float, and API responses keep rendering
these fields as JSON numbers via a single shared serializer, so no client
sees a wire-format change.

+++ app/providers/account_info.py
@@ -1,17 +1,24 @@
 import abc
+from decimal import Decimal
 from typing import NamedTuple
 
 
 # Typed read-only account snapshot shared by all account-info providers.
 # Mirrors MT5's account_info() fields at the application boundary so callers
 # never depend on the MetaTrader5 package directly.
+#
+# Money fields are Decimal (Step 37): balances, equity and margin are exact
+# financial quantities, converted at the provider boundary with
+# Decimal(str(raw_value)) so no binary-float artifact enters the domain.
+# margin_level is a ratio (equity / used margin * 100), not money, and stays
+# float.
 class AccountInfo(NamedTuple):
     login: int
     name: str
-    balance: float
-    equity: float
-    margin: float
-    free_margin: float
+    balance: Decimal
+    equity: Decimal
+    margin: Decimal
+    free_margin: Decimal
     margin_level: float
     currency: str
     server: str
