+++ app/providers/position.py
@@ -1,4 +1,5 @@
 import abc
+from decimal import Decimal
 from enum import StrEnum
 from typing import NamedTuple
 
@@ -13,14 +14,18 @@ class PositionType(StrEnum):
 # Typed read-only open-position snapshot shared by all position providers.
 # Explicit NamedTuple like Candle/AccountInfo so the raw MT5 object never
 # travels past the provider boundary.
+#
+# volume and the price/profit fields are Decimal (Step 37): volume is a
+# fixed-0.01-step quantity and open/current price and profit are financial
+# values, all converted at the provider boundary with Decimal(str(raw_value)).
 class Position(NamedTuple):
     ticket: int
     symbol: str
     type: PositionType
-    volume: float
-    open_price: float
-    current_price: float
-    profit: float
+    volume: Decimal
+    open_price: Decimal
+    current_price: Decimal
+    profit: Decimal
 
 
 # Abstraction boundary: services depend on this, never on MT5 directly
