+++ app/providers/trade_history.py
@@ -1,5 +1,6 @@
 import abc
 from datetime import datetime
+from decimal import Decimal
 from enum import StrEnum
 from typing import NamedTuple
 
@@ -28,23 +29,26 @@ class TradeCloseReason(StrEnum):
 # close_reason: DealCloseReason | None — None means MT5 supplied no reason for
 # this deal; the application must not guess one.
 #
-# stop_loss / take_profit: float | None. MT5's historical deal object does not
-# reliably carry the protective levels of the closing order, so they are only
-# filled when the related closing order can actually be retrieved
+# stop_loss / take_profit: Decimal | None. MT5's historical deal object does
+# not reliably carry the protective levels of the closing order, so they are
+# only filled when the related closing order can actually be retrieved
 # (history_orders_get(ticket=...)); otherwise they are None rather than
 # invented values.
+#
+# volume/price/profit and the protective levels are Decimal (Step 37),
+# converted at the provider boundary with Decimal(str(raw_value)).
 class TradeHistoryEntry(NamedTuple):
     ticket: int
     order_ticket: int
     symbol: str
     type: TradeType
-    volume: float
-    price: float
-    profit: float
+    volume: Decimal
+    price: Decimal
+    profit: Decimal
     time: datetime
     close_reason: TradeCloseReason | None
-    stop_loss: float | None
-    take_profit: float | None
+    stop_loss: Decimal | None
+    take_profit: Decimal | None
 
 
 # Abstraction boundary: services depend on this, never on MT5 directly
